bonitos のCast をdoubleからintへ

+++ src/jp/kobe_u/cspiral/norakore/model/User.java
@@ -45,8 +45,9 @@ public class User {
     public static int addBonitos(DBObject user, int new_bonitos) throws Exception {
         Object bo = user.get("bonitos");
         if (bo == null) throw new Exception("user's bonitos doesn't exist on DB.");
-        Double bonitos_d = (Double)bo;
-        int bonitos = bonitos_d.intValue() + new_bonitos;
+        //Double bonitos_d = (Double)bo;
+        int bonitos_d = (int)bo;
+        int bonitos = bonitos_d + new_bonitos;
         user.put("bonitos", bonitos);
         return bonitos;
     }
